LayoutTests/inspector/debugger/evaluateOnCallFrame-emulateUserGesture-hasTransientActivation-expected.txt

Tests for Debugger.evaluateOnCallFrame emulateUserGesture option.


== Running test suite: Debugger.evaluateOnCallFrame
-- Running test case: Debugger.evaluateOnCallFrame.emulateUserGesture.hasTransientActivation.false
Pausing...
before debugger
No transient activation
Resuming...
after debugger

-- Running test case: Debugger.evaluateOnCallFrame.emulateUserGesture.hasTransientActivation.true
Pausing...
before debugger
Has transient activation
Resuming...
after debugger
